Condense the `WorkflowRunner` sub-classes
The workflow output format should now be the same for both TF and Torch, so we can condense their respective sub-classes into the base `WorkflowRunner`, which only leaves the HugeCTR sub-class to deal with later (when we have HugeCTR support.)
Move `_convert_to_np` to the HugeCTR `WorkflowRunner`
